Tighten detection regexp to reject more non-svg input
The existing regexp matched any string with an occurrence of an <svg> tag
anywhere in the document.

By looking for a pair of tags and anchoring to the beginning and end of the
buffer, we can avoid a lot of false positives, including the README of this
module and the module's own source code.